+++ src/App.jsx
@@ -1,7 +1,6 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { useEffect } from "react";
 import "./App.css";
-import Lenis from "@studio-freight/lenis";
+import useLenisScroll from "./hooks/useLenisScroll";
 import Layout from "./components/Layout";
 import Home from "./pages/Home";
 import MovieDetail from "./pages/MovieDetail";
@@ -9,21 +8,7 @@ import Search from "./pages/Search";
 import Login from "./pages/Login";
 
 function App() {
-  useEffect(() => {
-    const lenis = new Lenis({
-      duration: 1.2,
-      easing: (t) => Math.min(1, 1.001 - Math.pow(2, -10 * t)),
-    });
-
-    function raf(time) {
-      lenis.raf(time);
-      requestAnimationFrame(raf);
-    }
-
-    requestAnimationFrame(raf);
-
-    return () => lenis.destroy();
-  });
+  useLenisScroll();
 
   return (
     <BrowserRouter>
