+++ src/hooks/useLenisScroll.js
@@ -1,0 +1,20 @@
+import { useEffect } from "react";
+import Lenis from "@studio-freight/lenis";
+
+export default function useLenisScroll() {
+  useEffect(() => {
+    const lenis = new Lenis({
+      duration: 1.2,
+      easing: (t) => Math.min(1, 1.001 - Math.pow(2, -10 * t)),
+    });
+
+    function raf(time) {
+      lenis.raf(time);
+      requestAnimationFrame(raf);
+    }
+
+    requestAnimationFrame(raf);
+
+    return () => lenis.destroy();
+  });
+}
